+++ src/libcore/data/mod.rs
@@ -1,0 +1,6 @@
+// src/libcore/data/mod.rs
+//
+// The data module contains modules for working with data-types, encoding data, decoding data, and using other programming-languages.
+
+// UCS-2
+pub mod ucs2;