+++ src/libcore/user/date.rs
@@ -1,0 +1,9 @@
+// src/libcore/user/date.rs
+//
+// The date command prints the current date.
+
+/*
+	IMPORTS
+*/
+
+use crate::time::{validate_fmtstr};