+++ src/libcore/user/mod.rs
@@ -1,6 +1,11 @@
 // src/libcore/user/mod.rs
 //
 // This is the mod.rs file for the libcore::user module.
+// The libcore::user module contains the basic functionality required for having users, providing commands, and providing the user with a shell.
 
 
+// The shell
 pub mod shell;
+
+// The ut command, which returns the time the system has gone without being restarted.
+pub mod ut;